feat: remove EntityBox from assets, segments and transactions
