+++ src/app/(routes)/segments/segments-data-table.tsx
@@ -128,13 +128,9 @@ export const SegmentsDataTable: React.FC<SegmentsTableProps> = (props) => {
 
   return (
     <FormProvider {...form}>
-      <EntityBox.Root>
-        <EntityBox.Actions className="flex w-full justify-end gap-4">
-          <div className="col-start-3 flex justify-end">
-            <PaginationLimitField control={form.control} />
-          </div>
-        </EntityBox.Actions>
-      </EntityBox.Root>
+      <div className="mb-4 flex justify-end">
+        <PaginationLimitField control={form.control} />
+      </div>
 
       <EntityDataTable.Root>
         {isNil(segments?.items) || segments.items.length === 0 ? (
